Fix dropdown overflow position being calculated too early
Clarity sets the dropdown anchor to `position: static` after dropdown is opened once, which means if we open the dropdown a second time, the next `position: relative` parent will be different. This is also a problem because the calculation when using `AfterViewChecked` considers the element to be at "y" of the anchor, not the actual "y" coordinate.

+ add an Input to configure the bottom margin of the dropdown