Upgade to Angular & Clarity 17
Angular did replace some @ with an escape sequence, because
the @ is reserved now (for the new Angular Control Flow).

A call to the rem() function in SCSS was fixed with the correct call.

Fixed two tests, that did return the translation with a leading line feed,
which shouldn't be a problem.