Angular & Clarity 17 (#2128)
* Upgade to Angular & Clarity 17

Angular did replace some @ with an escape sequence, because
the @ is reserved now (for the new Angular Control Flow).

A call to the rem() function in SCSS was fixed with the correct call.

Fixed two tests, that did return the translation with a leading line feed,
which shouldn't be a problem.

* Fix styles and components for Clarity 17

This commit updates the Clarity styles and components. All changes were carefully
tested with the development and demo pages. The changes include:

* Updating variable names for colors to align with Clarity Tokens
* Removing style overrides to align with Clarity
* Fixing issues with notifications and alerts
* Updating calendar styling and variable definitions
* Fixing dark theme compatibility
* Updating the Colors page
* Making minor fixes to spacing, alignment, and font sizes
* Making adaptations to the branding, spinner, and buttons
* Introducing a change log to track updates and breaking changes
* These changes ensure that the Clarity styles and components are up to date and
  consistent with the latest version.

A more detailed documentation of the changes can be found on the new
change-log page of the website.

* Fixes for Clarity 17

* Fixed the highlighting of active nav items.
* Fixed the hover color for dropdowns.
* Fixed the alignment of card headers.
* Fixed the margin of collapse-expand-sections.
* Reduced the contract of disabled buttons.
* Updated the package-lock once again.
* Removed commented out code.

* Distinguish between primary and info colors

Despite the fact that Clarity does not distinguish between primary
and info colors, this feature had been implemented for the Clarity Addons.

Due to the massive changes with Clarity 17, the support for this
feature was broken. Rebuilt the separation and made sure that
default themes still work.

* Replace "required" with "clr-required-mark"

Clarity is supporting this, now.

* Removed change log

In order to avoid repetitive work.