Minor fixed after the Clarity 17 upgrade (#2140)
* Improve visibility of search box

On the bright header, the search box wasn't visible at all. Added
the typical underline for input fields to the search box, because
the box wasn't easily detectable and in order ot visibly connect
the optional X on the right to the looking glass on the left.

* Fix the color of active cards

Previously, the primary color was used here.

* Fix the color of active links

The definition for the active color of links was missing.
And all links should be equally sized (not just the ones
with an href attrbitue).

* Fix nav menu in mobile mode

Clarity did change some settings that weren't
compatible with the main-nav-group.

* Update NG to serve https

Since most modern browsers redirect to https by default,
SSL has been activated for "ng serve".